Move bucket name generation to package object
