+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/gcs/GcsModel.scala
@@ -2,7 +2,6 @@ package org.broadinstitute.dsde.workbench.google.gcs
 
 import com.google.common.net.UrlEscapers
 import java.net.URI
-import java.util.UUID
 import org.broadinstitute.dsde.workbench.google.gcs.GcsPathParser._
 import scala.util.Try
 
@@ -20,39 +19,6 @@ case class GcsRelativePath(name: String) extends AnyVal
 
 /** A valid GCS bucket name */
 case class GcsBucketName(name: String) extends AnyVal
-object GcsBucketName {
-  /**
-    * Generates a unique bucket name with the given prefix. The prefix may be
-    * modified so that it adheres to bucket naming rules specified here:
-    *
-    * https://cloud.google.com/storage/docs/naming.
-    *
-    * The resulting bucket name is guaranteed to be a valid bucket name.
-    *
-    * @param prefix bucket name prefix
-    * @return generated bucket name
-    */
-  def generateUniqueBucketName(prefix: String) = {
-    // may only contain lowercase letters, numbers, underscores, dashes, or dots
-    val lowerCaseName = prefix.toLowerCase.filter { c =>
-      Character.isLetterOrDigit(c) || c == '_' || c == '-' || c == '.'
-    }
-
-    // must start with a letter or number
-    val sb = new StringBuilder(lowerCaseName)
-    if (!Character.isLetterOrDigit(sb.head)) sb.setCharAt(0, '0')
-
-    // max length of 63 chars, including the uuid
-    val uuid = UUID.randomUUID.toString
-    val maxNameLength = 63 - uuid.length - 1
-    if (sb.length > maxNameLength) sb.setLength(maxNameLength)
-
-    // must not start with "goog" or contain the string "google"
-    val processedName = sb.replaceAllLiterally("goog", "g00g")
-
-    GcsBucketName(s"$processedName-$uuid")
-  }
-}
 
 case class GcsParseError(message: String) extends AnyVal
